Creating 'completionHandlerAsync' attribute
This patch adds the core bits of the 'completionHandlerAsync' and gets
the compiler building again. Trying to use the attribute will cause
explosions.

This attribute will be used by tools to make refactoring and cleanup
recommendations, to move folks from using the older
completion handler-based code over to using the new async bits.