Computed properties can't be async, don't even try
The fix-it was being a little too eager to label function decls async.
Computed properties, or AccessorDecls are function declarations, but
they don't have parentheses, so trying to add `async` to them causes
crashing. Furthermore, they can't be async at all, so suggesting that we
make them async is just wrong. We shouldn't give folks a glimmer of hope
where there is none to be had.