Allow function references
This patch updates the typechecker to allow the function references
where the function is protected by an actor, so long as they aren't
called.

I still need a mechanism to ensure that the function reference isn't
passed to another function where it gets called. We can't really track
it through the other called function though.