Group closure suggestion
This patch coalesces all of capture-list errors into a single error.
This is necessary for IDE's to apply the fix-it appropriately, since it
won't update the fix-it list between fix-it applications. If we were
inserting a new capture list, this would end up fixing the code with
`{ [foo][bar][baz] in ... }` which isn't correct. Instead, we need to
insert everything that's missing in one fix-it.

Unfortunately, combining the fix-it meant that having multiple errors
didn't really make sense since the fix-it applies to all of them.