Run first thunk of Async main synchronously
This patch updates the asynchronous main function to run the first part
of the function synchronously.

The runloop is killed by exiting or aborting the task. As such, we need
to ensure that the task contains an async function that either calls
exit or aborts.

This means that there are now four layers of main. :)

@main: This is still the actual entry-point of the program. This
constructs the task containing @async_main, grabs the main executor,
runs swift_job_run to run the first part synchronously, and finally
kicks off the runloop with a call to _asyncMainDrainQueue. This is
generated through the call to emitAsyncMainThreadStart.

@async_main: This ensures that the operation passed into the task exits
or aborts rather than returning. It also consumes the error rather than
passing it down to @main. This actually calls out the normal
emitArtificialTopLevel with a few modifications for emitting a call to
'exit' instead of returning when working with an async function.

```
func async_main() async -> () {
  do {
    try await Main.$main()
    exit(0)
  } catch {
    _errorInMain(error)
  }
}
```

Main.$main(): This still has the same behaviour as with the synchronous
case. It just calls `try await Main.main()`.

Main.main(): This is the same as in the synchronous case, but async.
This is the user-specified code.