Do force casts instead of dyn casts
I had used asserts followed by dyn_casts to destabilizing the release
compiler when checking for the 'await' in an interpolated string
expression. The asserts should have caught something on the asserts
builds had it been an issue. I haven't been able to come up with an
expression where the interpolated string isn't a call expression
containing a paren expression.

This is to cleanup this code and avoid doing who-knows-what in the event
that my assumptions are not holding in release builds.