Handle errors for async closures
Had to do a little design here. We can emit async closures from
synchronous functions. Those should still get checked.
As a result, keeping a stack of closures that we're in allows us to do
the appropriate checking to verify that we're in a nested async closure
before emitting the diagnostic.