Fix main resolution
When I tried to make the type system allow async main functions, I
didn't take into account `where` clauses on different extensions.
This patch removes what I had done before and wires the logic for
permitting async main functions into the constraint system resolver
directly.

The new resolution behavior looks at whether the function is an async
main function. If it isn't, and the targeted version of swift permits
using concurrency, the score for that overload is incremented, giving
preference to the async version, but still adhering to things like
where-clauses.

The attached test ensures that we resolve to the correct main function
when there is a "default" that is async, a more specified synchronous
main, or when there is a more-specified async main.