Importing stuff
The tool is starting to import data into the database now.
At this point, we can collect enough information to start recording file
sizes. This means that we have info about the platform, build, and the
files getting collected. I am not yet collecting enough info about the
directory structure yet though, so we can't aggregate or count the
number of files in a given directory. That will be next and will
probably involve some fun recursive queries!