Add directories
Adding directories to the files and sizes list. This breaks the
normalization a bit and restricts what we can do, but gets the tool up
and running. It would be nice to break the path down into a file
hierarchy represented in the database, but that's also more work and
leads to more complicated queries to get summary information.