Plumb correct swiftc to swiftpm
The swiftpm was deriving the swiftc from the native tools directory for
the product, which was in turn derived from the swiftc path according to
either `native_swift_tools_path` or from the install location.

The new logic is if `native_swift_tools_path` is set, use that as the
native path. Otherwise, if the toolchain has a swift compiler, use the
path two levels up from that.

Finally, if the toolchain doesn't include a swiftc, then fall back on
the install location.

Now, it's important to note that toolchain objects don't inherently have
a single directory. MacOS uses `xcrun`, Linux and Windows use `which`,
and the custom toolchain does a filesystem walk to find the tool. In the
event that `swiftc` is not located at '<root>/bin/swiftc', the
heuristic in `products.native_toolchain_path` will fail to produce the
correct toolchain root.