WIP: Add diagnostics for missing entrypoints
This patch adds two flags to control emitting diagnostics for
entrypoints. One is to ensure that the module does not contain an
entrypoint, and the other is to ensure that it does.

Note that the `main` file is whichever file comes first, so an
entrypoint may appear in a non-main file.

This is to provide some diagnostics when the build-system knows whether
or not it is building a library or executable, so that we can provide a
few helpful diagnostics. So as to avoid breaking compatibility, the
default state is to ignore these diagnostics altogether. Only if one of
`-verify-no-entrypoint` or `-verify-entrypoint` are set do we diagnose
the presence of an entry point or a missing entrypoint, respectively.