Build stdlib with correct CMake usage
This patch migrates most of the standard library and supporting
libraries to build with correct CMake. This should make it easier to
maintain, cross-compile the standard library, and work with.

We need to isolate the Swift shims built in the stdlib build from the
shims that already exist in the SDK In order to do this, we need
explicit module builds. Explicit modules builds are a bit stricter about
what they do and don't allow, so I've gone through and added the
specific libraries that things are depending on.